address duplicated code from sonarcloud
